% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KoboldR_function.R
\name{Generate_code}
\alias{Generate_code}
\title{A Code Instruction Function}
\usage{
Generate_code(Prompt)
}
\arguments{
\item{Prompt}{The input text given to the model to generate code instruction.}
}
\description{
This function is specifically tuned for code instruction in R.
}
\examples{
program <- 'Could you write a piece of R code that defines a function, which gives the mean and sd of a data frame?'
Generate_code(program)
}
\keyword{code,}
\keyword{instruction}
\keyword{text,}
