% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KoboldR_function.R
\name{Generate_text}
\alias{Generate_text}
\title{A Text Generation Function}
\usage{
Generate_text(Prompt, Memory, Genkey = "KCPP2531")
}
\arguments{
\item{Prompt}{The input text given to the model to generate a response.}

\item{Memory}{Context given to the model in text generation.}

\item{Genkey}{A unique genkey of each generation session, used in multiple session scenario. Default to "KCPP2531".}
}
\description{
This function allows you to generate text given a prompt and generation settings.
}
\examples{
greeting <- "Hi there"
Generate_text(greeting, Memory = '') |> cat()
}
\keyword{generation}
\keyword{text,}
