% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KoboldR_function.R
\name{Image_generate_text}
\alias{Image_generate_text}
\title{An Image-to-Text Generation Function}
\usage{
Image_generate_text(Prompt, Memory, Image_path, Genkey = "KCPP2531")
}
\arguments{
\item{Prompt}{The input text given to the model to generate a response.}

\item{Memory}{Context given to the model in text generation.}

\item{Image_path}{file path of the input image}

\item{Genkey}{A unique genkey of each generation session, used in multiple session scenario. Default to "KCPP2531".}
}
\description{
This function allows you to generate text given a prompt, an image, and generation settings.
}
\examples{
Image_generate_text(
          'Could you describe the image?', 
          Memory = '', 
          'Cats_purring.png'
          ) |> cat()
}
\keyword{generation}
\keyword{image,}
\keyword{text,}
